lets speed up

+++ CFR/Node.cpp
@@ -5,13 +5,13 @@
 namespace CFR {
 /// @param actionNum Number of available actions in this node
     Node::Node(const int actionNum) : mActionNum(actionNum), mAlreadyCalculated(false), mNeedToUpdateStrategy(false) {
-        mRegretSum = new double[actionNum];
-        mStrategy = new double[actionNum];
-        mStrategySum = new double[actionNum];
-        mAverageStrategy = new double[actionNum];
+        mRegretSum = new float[actionNum];
+        mStrategy = new float[actionNum];
+        mStrategySum = new float[actionNum];
+        mAverageStrategy = new float[actionNum];
         for (int a = 0; a < actionNum; ++a) {
             mRegretSum[a] = 0.0;
-            mStrategy[a] = 1.0 / (double) actionNum;
+            mStrategy[a] = 1.0 / (float) actionNum;
             mStrategySum[a] = 0.0;
             mAverageStrategy[a] = 0.0;
         }
@@ -26,13 +26,13 @@ namespace CFR {
 
 /// @brief Get the current getStrategy at this node through Regret-Matching
 /// @return mixed getStrategy
-    const double *Node::getStrategy() {
+    const float *Node::getStrategy() {
         return mStrategy;
     }
 
 /// @brief Get the average getStrategy across all training iterations
 /// @return average mixed getStrategy
-    const double *Node::averageStrategy() {
+    const float *Node::averageStrategy() {
         if (!mAlreadyCalculated) {
             calcAverageStrategy();
         }
@@ -43,7 +43,7 @@ namespace CFR {
 ///        The contribution is the probability of reaching this node if all players other than the acting player always choose actions leading to this node.
 /// @param strategy current getStrategy
 /// @param realizationWeight contribution of the acting player at this node
-    void Node::strategySum(const double *strategy, const double realizationWeight) {
+    void Node::strategySum(const float *strategy, const float realizationWeight) {
         for (int a = 0; a < mActionNum; ++a) {
             mStrategySum[a] += realizationWeight * strategy[a];
         }
@@ -55,7 +55,7 @@ namespace CFR {
         if (!mNeedToUpdateStrategy) {
             return;
         }
-        double normalizingSum = 0.0;
+        float normalizingSum = 0.0;
         for (int a = 0; a < mActionNum; ++a) {
             mStrategy[a] = mRegretSum[a] > 0 ? mRegretSum[a] : 0;
             normalizingSum += mStrategy[a];
@@ -64,7 +64,7 @@ namespace CFR {
             if (normalizingSum > 0) {
                 mStrategy[a] /= normalizingSum;
             } else {
-                mStrategy[a] = 1.0 / (double) mActionNum;
+                mStrategy[a] = 1.0 / (float) mActionNum;
             }
         }
     }
@@ -72,14 +72,14 @@ namespace CFR {
 /// @brief Get the cumulative counterfactual regret of the specified action
 /// @param action action
 /// @return cumulative counterfactual regret
-    double Node::regretSum(const int action) const {
+    float Node::regretSum(const int action) const {
         return mRegretSum[action];
     }
 
 /// @brief Update the cumulative counterfactual regret by doing addition the counterfactual regret weighted by the contribution of the all players other than the acting player at this node.
 /// @param action action
 /// @param value counterfactual regret
-    void Node::regretSum(const int action, const double value) {
+    void Node::regretSum(const int action, const float value) {
         mRegretSum[action] = value;
         mNeedToUpdateStrategy = true;
     }
@@ -101,7 +101,7 @@ namespace CFR {
         for (int a = 0; a < mActionNum; ++a) {
             mAverageStrategy[a] = 0.0;
         }
-        double normalizingSum = 0.0;
+        float normalizingSum = 0.0;
         for (int a = 0; a < mActionNum; ++a) {
             normalizingSum += mStrategySum[a];
         }
@@ -109,7 +109,7 @@ namespace CFR {
             if (normalizingSum > 0) {
                 mAverageStrategy[a] = mStrategySum[a] / normalizingSum;
             } else {
-                mAverageStrategy[a] = 1.0 / (double) mActionNum;
+                mAverageStrategy[a] = 1.0 / (float) mActionNum;
             }
         }
         mAlreadyCalculated = true;
