+++ CFR/Node.hpp
@@ -18,17 +18,17 @@ namespace CFR {
 
         /// @brief Get the current getStrategy at this node through Regret-Matching
         /// @return mixed getStrategy
-        const double *getStrategy();
+        const float *getStrategy();
 
         /// @brief Get the average getStrategy across all training iterations
         /// @return average mixed getStrategy
-        const double *averageStrategy();
+        const float *averageStrategy();
 
         /// @brief Update the average strategy by doing addition the current getStrategy weighted by the contribution of the acting player at the this node.
         ///        The contribution is the probability of reaching this node if all players other than the acting player always choose actions leading to this node.
         /// @param strategy current getStrategy
         /// @param realizationWeight contribution of the acting player at this node
-        void strategySum(const double *strategy, double realizationWeight);
+        void strategySum(const float *strategy, float realizationWeight);
 
         /// @brief Update current getStrategy
         void updateStrategy();
@@ -36,12 +36,12 @@ namespace CFR {
         /// @brief Get the cumulative counterfactual regret of the specified action
         /// @param action action
         /// @return cumulative counterfactual regret
-        double regretSum(int action) const;
+        float regretSum(int action) const;
 
         /// @brief Update the cumulative counterfactual regret by doing addition the counterfactual regret weighted by the contribution of the all players other than the acting player at this node.
         /// @param action action
         /// @param value counterfactual regret
-        void regretSum(int action, double value);
+        void regretSum(int action, float value);
 
         /// @brief Get the number of available actions at this node.
         /// @return number of available actions
@@ -55,10 +55,10 @@ namespace CFR {
 
 
         uint8_t mActionNum;
-        double *mRegretSum;
-        double *mStrategy;
-        double *mStrategySum;
-        double *mAverageStrategy;
+        float *mRegretSum;
+        float *mStrategy;
+        float *mStrategySum;
+        float *mAverageStrategy;
         bool mAlreadyCalculated;
         bool mNeedToUpdateStrategy;
     };
